fix: rename conflicting WeatherJsonContext (root cause of P0-1 generator block)
CS8785 revealed the true cause: the new WeatherHandlers.cs declared
WeatherJsonContext, colliding with the existing partial context in
WeatherService.cs. Duplicate hintName made the STJ generator abandon all
output for the assembly (160 cascade CS0534). One-line rename restores
the build. The Huorong-block hypothesis is disproven; known-issues doc
updated to resolved state with root cause and lesson learned.

+++ src/DeskBox/Services/CommandApi/Handlers/GlanceHandlers.cs
@@ -16,7 +16,7 @@ public sealed record GlanceInfoResult(
     string Layout,
     string Transition,
     int LocalImageCount,
-    int RotationIntervalMinutes,
+    double RotationIntervalMinutes,
     bool RandomOrder);
 
 [JsonSourceGenerationOptions(
